requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
gradio>=5.0.0
python-dotenv>=1.0.0
ordec>=0.3.0
pytest>=8.0.0
